+++ GodotTestDriver/Input/KeyboardControlExtensions.cs
@@ -14,16 +14,30 @@ namespace GodotTestDriver.Input
         /// <summary>
         /// Presses the given key with the given modifiers.
         /// </summary>
-        public static async Task PressKey(this Node node, Key key, bool control = false, bool alt = false,
-            bool shift = false, bool meta = false)
+        /// <param name="key">Keyboard key.</param>
+        /// <param name="control">Control modifier.</param>
+        /// <param name="alt">Alt modifier.</param>
+        /// <param name="shift">Shift modifier.</param>
+        /// <param name="meta">Meta-key (Windows, Command) modifier.</param>
+        /// <returns>Asynchronous task completed when key is released.</returns>
+        public static async Task PressKey(
+            this Node node,
+            Key key,
+            bool control = false,
+            bool alt = false,
+            bool shift = false,
+            bool meta = false
+        )
         {
-            var inputEvent = new InputEventKey();
-            inputEvent.Pressed = true;
-            inputEvent.Keycode = key;
-            inputEvent.AltPressed = alt;
-            inputEvent.CtrlPressed = control;
-            inputEvent.ShiftPressed = shift;
-            inputEvent.MetaPressed = meta;
+            var inputEvent = new InputEventKey
+            {
+                Pressed = true,
+                Keycode = key,
+                AltPressed = alt,
+                CtrlPressed = control,
+                ShiftPressed = shift,
+                MetaPressed = meta
+            };
 
             Godot.Input.ParseInputEvent(inputEvent);
 
@@ -31,18 +45,57 @@ namespace GodotTestDriver.Input
         }
 
         /// <summary>
+        /// Simulate a key being pressed for a certain amount of time.
+        /// </summary>
+        /// <param name="milliseconds">How long the key should be held for.</param>
+        /// <param name="key">Keyboard key.</param>
+        /// <param name="control">Control modifier.</param>
+        /// <param name="alt">Alt modifier.</param>
+        /// <param name="shift">Shift modifier.</param>
+        /// <param name="meta">Meta-key (Windows, Command) modifier.</param>
+        /// <returns>Asynchronous task completed when key is released.</returns>
+        public static async Task HoldKeyFor(
+            this Node node,
+            float seconds,
+            Key key,
+            bool control = false,
+            bool alt = false,
+            bool shift = false,
+            bool meta = false
+        )
+        {
+            await node.PressKey(key, control, alt, shift, meta);
+            await node.Wait(seconds);
+            await node.ReleaseKey(key, control, alt, shift, meta);
+        }
+
+        /// <summary>
         /// Releases the given key with the given modifier state.
         /// </summary>
-        public static async Task ReleaseKey(this Node node, Key key, bool control = false, bool alt = false,
-            bool shift = false, bool meta = false)
+        /// <param name="key">Keyboard key.</param>
+        /// <param name="control">Control modifier.</param>
+        /// <param name="alt">Alt modifier.</param>
+        /// <param name="shift">Shift modifier.</param>
+        /// <param name="meta">Meta-key (Windows, Command) modifier.</param>
+        /// <returns>Asynchronous task completed when key is released.</returns>
+        public static async Task ReleaseKey(
+            this Node node,
+            Key key,
+            bool control = false,
+            bool alt = false,
+            bool shift = false,
+            bool meta = false
+        )
         {
-            var inputEvent = new InputEventKey();
-            inputEvent.Pressed = false;
-            inputEvent.Keycode =  key;
-            inputEvent.CtrlPressed = control;
-            inputEvent.AltPressed = alt;
-            inputEvent.ShiftPressed = shift;
-            inputEvent.MetaPressed = meta;
+            var inputEvent = new InputEventKey
+            {
+                Pressed = false,
+                Keycode = key,
+                CtrlPressed = control,
+                AltPressed = alt,
+                ShiftPressed = shift,
+                MetaPressed = meta
+            };
 
             Godot.Input.ParseInputEvent(inputEvent);
 
@@ -52,8 +105,20 @@ namespace GodotTestDriver.Input
         /// <summary>
         /// Presses and releases a key with the given modifiers.
         /// </summary>
-        public static async Task TypeKey(this Node node, Key key, bool control = false, bool alt = false,
-            bool shift = false, bool meta = false)
+        /// <param name="key">Keyboard key.</param>
+        /// <param name="control">Control modifier.</param>
+        /// <param name="alt">Alt modifier.</param>
+        /// <param name="shift">Shift modifier.</param>
+        /// <param name="meta">Meta-key (Windows, Command) modifier.</param>
+        /// <returns>Asynchronous task completed when key is released.</returns>
+        public static async Task TypeKey(
+            this Node node,
+            Key key,
+            bool control = false,
+            bool alt = false,
+            bool shift = false,
+            bool meta = false
+        )
         {
             await node.PressKey(key, control, alt, shift, meta);
             await node.ReleaseKey(key, control, alt, shift, meta);